requirements.txt
python-dotenv==1.0.0
slack-sdk==3.26.1
slack-bolt==1.18.1
pytz==2024.1
requests==2.31.0 
